+++ crates/stackable-versioned/CHANGELOG.md
@@ -4,9 +4,12 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+- Change from derive macro to attribute macro to be able to generate code
+  _in place_ instead of _appending_ new code ([#CHANGEME]).
 - Improve action chain generation ([#784]).
 
-[#784](ttps://github.com/stackabletech/operator-rs/pull/784)
+[#784](https://github.com/stackabletech/operator-rs/pull/784)
+[#CHANGEME](https://github.com/stackabletech/operator-rs/pull/CHANGEME)
 
 - Improve action chain generation ([#784]).